Update lisk-codec to use lisk schema from lisk-validator
