+++ elements/lisk-codec/package.json
@@ -34,8 +34,7 @@
 		"prepublishOnly": "npm run lint && npm test && npm run build && npm run build:check"
 	},
 	"dependencies": {
-		"@liskhq/lisk-validator": "0.4.0-alpha.0",
-		"debug": "4.1.1"
+		"@liskhq/lisk-validator": "0.4.0-alpha.0"
 	},
 	"devDependencies": {
 		"@types/jest": "25.1.3",
@@ -52,7 +51,6 @@
 		"jest": "25.1.0",
 		"jest-extended": "0.11.5",
 		"jest-when": "2.7.0",
-		"lodash.clonedeep": "4.5.0",
 		"prettier": "1.19.1",
 		"source-map-support": "0.5.16",
 		"ts-jest": "25.2.1",
