+++ elements/lisk-codec/src/index.ts
@@ -13,4 +13,3 @@
  */
 
 export * from './codec';
-export { validateSchema } from './schema';
